push props down to SelectPage
